+++ components/form/input-select.tsx
@@ -65,9 +65,6 @@ const Select = styled('select', {
   },
 })
 
-
-
-
 export const InputSelect = ({ options, value, id, name, label, required, onChange }: InputProps) => {
   return (
     <InputBox>
@@ -81,7 +78,7 @@ export const InputSelect = ({ options, value, id, name, label, required, onChang
 
         {options.map((opt) => {
           return (
-            <Option id={opt.value} value={opt.value}>{opt.label}</Option>
+            <Option key={opt.value} value={opt.value}>{opt.label}</Option>
           )
         })}
       </Select>
